requirements.txt
streamlit
plotly.express
pandas
math